Adding Paxmesh
Since cosmos doesn't generate a new paxmesh if it sees an old one, any model updates would require the user to delete the old ones to make cosmos generate a new one.  Keeping the paxmesh in the repo will make it so that cosmos doesn't need to look, the model is already updated.